[OST]코드 모듈화
한 py파일에 있던 코드들 기능별로 py파일로 만들어 모듈화 한 후 main파일에서 호출하도록 변경함
