+++ .idea/.name
@@ -1,0 +1,1 @@
+stock_recom_chatbot.py
