+++ .idea/inspectionProfiles/Project_Default.xml
@@ -1,0 +1,28 @@
+<component name="InspectionProjectProfileManager">
+  <profile version="1.0">
+    <option name="myName" value="Project Default" />
+    <inspection_tool class="PyPackageRequirementsInspection" enabled="true" level="WARNING" enabled_by_default="true">
+      <option name="ignoredPackages">
+        <value>
+          <list size="15">
+            <item index="0" class="java.lang.String" itemvalue="chromadb" />
+            <item index="1" class="java.lang.String" itemvalue="tiktoken" />
+            <item index="2" class="java.lang.String" itemvalue="bs4" />
+            <item index="3" class="java.lang.String" itemvalue="faiss-cpu" />
+            <item index="4" class="java.lang.String" itemvalue="finance-datareader" />
+            <item index="5" class="java.lang.String" itemvalue="streamlit" />
+            <item index="6" class="java.lang.String" itemvalue="sentence-transformers" />
+            <item index="7" class="java.lang.String" itemvalue="mplfinance" />
+            <item index="8" class="java.lang.String" itemvalue="plotly" />
+            <item index="9" class="java.lang.String" itemvalue="langchain_community" />
+            <item index="10" class="java.lang.String" itemvalue="loguru" />
+            <item index="11" class="java.lang.String" itemvalue="openai" />
+            <item index="12" class="java.lang.String" itemvalue="langchain" />
+            <item index="13" class="java.lang.String" itemvalue="unstructured" />
+            <item index="14" class="java.lang.String" itemvalue="pysqlite3-binary" />
+          </list>
+        </value>
+      </option>
+    </inspection_tool>
+  </profile>
+</component>
