--- stock_recom_chatbot.py
@@ -1,389 +1,0 @@
-import streamlit as st
-import requests
-import urllib.parse
-import FinanceDataReader as fdr
-import tiktoken
-from bs4 import BeautifulSoup
-from datetime import datetime, timedelta
-from langchain.text_splitter import RecursiveCharacterTextSplitter
-from langchain.embeddings import HuggingFaceEmbeddings
-from langchain.vectorstores import FAISS
-from langchain.chat_models import ChatOpenAI
-from langchain.chains import ConversationalRetrievalChain
-from langchain.memory import ConversationBufferMemory
-import pandas as pd
-import yfinance as yf
-import plotly.graph_objects as go
-import os
-import random
-from difflib import SequenceMatcher
-import urllib.parse
-from sklearn.feature_extraction.text import TfidfVectorizer
-from sklearn.metrics.pairwise import cosine_similarity
-
-
-def main():
-    st.set_page_config(page_title="Stock Analysis Chatbot", page_icon=":chart_with_upwards_trend:")
-    st.title("기업 정보 분석 QA Chat")
-
-    if "conversation" not in st.session_state:
-        st.session_state.conversation = None
-    if "chat_history" not in st.session_state:
-        st.session_state.chat_history = None
-    if "processComplete" not in st.session_state:
-        st.session_state.processComplete = False
-    if "news_data" not in st.session_state:
-        st.session_state.news_data = None
-    if "company_name" not in st.session_state:
-        st.session_state.company_name = None
-    if "selected_period" not in st.session_state:
-        st.session_state.selected_period = "1day"
-
-
-    with st.sidebar:
-        openai_api_key = st.text_input("OpenAI API Key", key="chatbot_api_key", type="password")
-        company_name = st.text_input("분석할 기업명 (코스피 상장)")
-        days = st.number_input("최근 며칠 동안의 기사를 검색할까요?", min_value=1, max_value=30, value=7)  # 기간을 사용자 입력받기
-        process = st.button("분석 시작")
-
-    if process:
-        if not openai_api_key or not company_name:
-            st.info("OpenAI API 키와 기업명을 입력해주세요.")
-            st.stop()
-
-        news_data = crawl_news(company_name, days)
-        if not news_data:
-            st.warning("해당 기업의 최근 뉴스를 찾을 수 없습니다.")
-            st.stop()
-            
-
-        # 분석 결과를 session_state에 저장
-        st.session_state.news_data = news_data
-        st.session_state.company_name = company_name
-
-        text_chunks = get_text_chunks(news_data)
-        vectorstore = get_vectorstore(text_chunks)
-        
-        st.session_state.conversation = create_chat_chain(vectorstore, openai_api_key)
-        st.session_state.processComplete = True
-
-    # 분석 결과가 있으면 항상 상단에 출력
-    if st.session_state.processComplete and st.session_state.company_name:
-        st.subheader(f"📈 {st.session_state.company_name} 최근 주가 추이")
-
-        # ✅ 선택된 기간을 강제 업데이트하여 즉시 반영
-        st.session_state.radio_selection = st.session_state.selected_period
-        selected_period = st.radio(
-            "기간 선택",
-            options=["1day", "week", "1month", "1year"],
-            index=["1day", "week", "1month", "1year"].index(st.session_state.selected_period),
-            key="radio_selection",
-            on_change=update_period  # ✅ 선택 즉시 반영
-        )
-
-      
-        if selected_period != st.session_state.selected_period:
-            st.session_state.selected_period = selected_period
-
-
-        st.write(f"🔍 선택된 기간: {st.session_state.selected_period}")
-
-        with st.spinner(f"📊 {st.session_state.company_name} ({st.session_state.selected_period}) 데이터 불러오는 중..."):
-
-            if selected_period in ["1day", "week"]:
-                ticker = get_ticker(st.session_state.company_name, source="yahoo")  # ✅ 야후 파이낸스용 티커
-
-                if not ticker:
-                    st.error("해당 기업의 야후 파이낸스 티커 코드를 찾을 수 없습니다.")
-                    return
-
-                interval = "1m" if st.session_state.selected_period == "1day" else "5m"
-                df = get_intraday_data_yahoo(ticker,
-                                             period="5d" if st.session_state.selected_period == "week" else "1d",
-                                             interval=interval)
-
-            else:
-                ticker = get_ticker(st.session_state.company_name, source="fdr")  # ✅ FinanceDataReader용 티커
-                if not ticker:
-                    st.error("해당 기업의 FinanceDataReader 티커 코드를 찾을 수 없습니다.")
-                    return
-
-                df = get_daily_stock_data_fdr(ticker, st.session_state.selected_period)
-
-            if df.empty:
-                st.warning(
-                    f"📉 {st.session_state.company_name} - 해당 기간({st.session_state.selected_period})의 거래 데이터가 없습니다.")
-            else:
-                plot_stock_plotly(df, st.session_state.company_name, st.session_state.selected_period)
-
-        st.markdown("최근 기업 뉴스 목록을 보려면 누르시오")
-
-    if st.session_state.processComplete:
-        with st.expander("뉴스 보기"):
-            news_data = st.session_state.news_data
-
-            # 처음 10개 뉴스만 표시
-            for i, news in enumerate(news_data[:10]):
-                st.markdown(f"- **{news['title']}** ([링크]({news['link']}))")
-
-            # '더보기' 버튼 클릭 시 나머지 뉴스 표시
-            if len(news_data) > 10:
-                if st.button('더보기', key="show_more"):
-                    for news in news_data[10:]:
-                        st.markdown(f"- **{news['title']}** ([링크]({news['link']}))")
-
-
-    # 채팅 부분: 사용자가 질문을 입력하면 대화가 이어짐
-    if query := st.chat_input("질문을 입력해주세요."):
-        with st.chat_message("user"):
-            st.markdown(query)
-
-        with st.chat_message("assistant"):
-            with st.spinner("분석 중..."):
-                result = st.session_state.conversation({"question": query})
-                response = result['answer']
-
-                st.markdown(response)
-                with st.expander("참고 뉴스 확인"):
-                    for doc in result['source_documents']:
-                        st.markdown(f"- [{doc.metadata['source']}]({doc.metadata['source']})")
-
-
-
-def crawl_news(company, days, threshold=0.3):
-    today = datetime.today()
-    start_date = (today - timedelta(days=days)).strftime('%Y%m%d')
-    end_date = today.strftime('%Y%m%d')
-    encoded_query = urllib.parse.quote(company)
-
-    url_template = f"https://search.naver.com/search.naver?where=news&query={encoded_query}&nso=so:r,p:from{start_date}to{end_date}&start={{}}"
-
-    headers = {
-        "User-Agent": random.choice([
-            "Mozilla/5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/91.0.4472.124 Safari/537.36",
-            "Mozilla/5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/92.0.4515.159 Safari/537.36",
-            "Mozilla/5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/93.0.4577.82 Safari/537.36"
-        ])
-    }
-
-    data = []
-    for page in range(1, 6):
-        url = url_template.format((page - 1) * 10 + 1)
-        response = requests.get(url, headers=headers)
-        response.raise_for_status()
-        soup = BeautifulSoup(response.text, 'html.parser')
-        articles = soup.select("ul.list_news > li")
-
-        for article in articles:
-            title = article.select_one("a.news_tit").text
-            link = article.select_one("a.news_tit")['href']
-            content = article.select_one("div.news_dsc").text if article.select_one("div.news_dsc") else ""
-            data.append({"title": title, "link": link, "content": content})
-
-    return deduplicate_news(data, threshold)
-
-
-def deduplicate_news(news_data, threshold=0.3):
-    if len(news_data) <= 1:
-        return news_data
-
-    # 제목과 본문을 합친 텍스트 생성
-    combined_texts = [news['title'] + " " + news['content'] for news in news_data]
-    vectorizer = TfidfVectorizer().fit_transform(combined_texts)
-    cosine_sim = cosine_similarity(vectorizer, vectorizer)
-
-    filtered_news = []
-    seen_indices = set()
-
-    for i, news in enumerate(news_data):
-        if i in seen_indices:
-            continue
-
-        filtered_news.append(news)
-        for j in range(i + 1, len(news_data)):
-            if news_data[j]['title'] == news['title'] or cosine_sim[i, j] > threshold:
-                seen_indices.add(j)
-
-    return filtered_news
-
-
-def tiktoken_len(text):
-    tokenizer = tiktoken.get_encoding("cl100k_base")
-    tokens = tokenizer.encode(text)
-    return len(tokens)
-
-def get_text_chunks(news_data):
-    texts = [f"{item['title']}\n{item['content']}" for item in news_data]
-    metadatas = [{"source": item["link"]} for item in news_data]
-    text_splitter = RecursiveCharacterTextSplitter(
-        chunk_size=900,
-        chunk_overlap=100,
-        length_function=tiktoken_len
-    )
-    return text_splitter.create_documents(texts, metadatas=metadatas)
-
-def get_vectorstore(text_chunks):
-    embeddings = HuggingFaceEmbeddings(
-        model_name="jhgan/ko-sroberta-multitask",
-        model_kwargs={'device': 'cpu'},
-        encode_kwargs={'normalize_embeddings': True}
-    )
-    return FAISS.from_documents(text_chunks, embeddings)
-
-def create_chat_chain(vectorstore, openai_api_key):
-    llm = ChatOpenAI(openai_api_key=openai_api_key, model_name='gpt-4', temperature=0)
-    return ConversationalRetrievalChain.from_llm(
-        llm=llm, chain_type="stuff", retriever=vectorstore.as_retriever(),
-        memory=ConversationBufferMemory(memory_key='chat_history', return_messages=True, output_key='answer'),
-        get_chat_history=lambda h: h, return_source_documents=True)
-
-
-# ✅ 세션 상태 업데이트 함수 (기간 변경 시 즉시 반영)
-def update_period():
-    st.session_state.selected_period = st.session_state.radio_selection
-
-
-
-# ✅ 1. 최근 거래일 찾기 함수
-def get_recent_trading_day():
-    today = datetime.now()
-    if today.hour < 9:  # 9시 이전이면 전날을 기준으로
-        today -= timedelta(days=1)
-    while today.weekday() in [5, 6]:  # 토요일(5), 일요일(6)이면 하루씩 감소
-        today -= timedelta(days=1)
-    return today.strftime('%Y-%m-%d')
-
-# ✅ 2. 티커 조회 함수 (야후 & FinanceDataReader)
-def get_ticker(company, source="yahoo"):
-    try:
-        listing = fdr.StockListing('KRX')
-        ticker_row = listing[listing["Name"].str.strip() == company.strip()]
-        if not ticker_row.empty:
-            krx_ticker = str(ticker_row.iloc[0]["Code"]).zfill(6)
-            if source == "yahoo":
-                return krx_ticker + ".KS"  # ✅ 야후 파이낸스용 티커 변환
-            return krx_ticker  # ✅ FinanceDataReader용 티커
-        return None
-    except Exception as e:
-        st.error(f"티커 조회 중 오류 발생: {e}")
-        return None
-
-# ✅ 3. 야후 파이낸스에서 분봉 데이터 가져오기 (1day, week)
-def get_intraday_data_yahoo(ticker, period="1d", interval="1m"):
-    try:
-        stock = yf.Ticker(ticker)
-        df = stock.history(period=period, interval=interval)
-        if df.empty:
-            return pd.DataFrame()
-        df = df.reset_index()
-        df = df.rename(columns={"Datetime": "Date", "Close": "Close",
-                                "Open": "Open", "High": "High", "Low": "Low"})
-        # ✅ 주말 데이터 제거 (혹시 남아있는 경우 대비)
-        df["Date"] = pd.to_datetime(df["Date"])
-        df = df[df["Date"].dt.weekday < 5].reset_index(drop=True)
-        return df
-    except Exception as e:
-        st.error(f"야후 파이낸스 데이터 불러오기 오류: {e}")
-        return pd.DataFrame()
-
-# ✅ 4. FinanceDataReader를 통한 일별 시세 (1month, 1year)
-def get_daily_stock_data_fdr(ticker, period):
-    try:
-        end_date = get_recent_trading_day()
-        start_date = (datetime.strptime(end_date, '%Y-%m-%d') - timedelta(days=30 if period == "1month" else 365)).strftime('%Y-%m-%d')
-        df = fdr.DataReader(ticker, start_date, end_date)
-        if df.empty:
-            return pd.DataFrame()
-        df = df.reset_index()
-        df = df.rename(columns={"Date": "Date", "Close": "Close"})
-        # ✅ 주말 데이터 완전 제거
-        df["Date"] = pd.to_datetime(df["Date"])
-        df = df[df["Date"].dt.weekday < 5].reset_index(drop=True)
-        return df
-    except Exception as e:
-        st.error(f"FinanceDataReader 데이터 불러오기 오류: {e}")
-        return pd.DataFrame()
-
-
-# ✅ 5. Plotly를 이용한 주가 시각화 함수 (x축 포맷 최적화)
-
-def plot_stock_plotly(df, company, period):
-    if df is None or df.empty:
-        st.warning(f"📉 {company} - 해당 기간({period})의 거래 데이터가 없습니다.")
-        return
-
-    fig = go.Figure()
-
-    # ✅ x축 날짜 형식 설정
-    if period == "1day":
-        df["FormattedDate"] = df["Date"].dt.strftime("%H:%M")
-    elif period == "week":
-        df["FormattedDate"] = df["Date"].dt.strftime("%m-%d %H:%M")
-    else:
-        df["FormattedDate"] = df["Date"].dt.strftime("%m-%d")
-
-    # ✅ x축 간격 설정
-    if period == "1day":
-        tickvals = df.iloc[::60]["FormattedDate"].tolist()  # 1시간 간격
-    elif period == "week":
-        tickvals = df[df["FormattedDate"].str.endswith("09:00")]["FormattedDate"].tolist()  # 9시만 표시
-    elif period == "1month":
-        tickvals = df.iloc[::4]["FormattedDate"].tolist()  # 4일 간격
-    else:  # 1year - 첫 달은 건너뛰고 나머지 월만 표시
-        df['Year'] = df['Date'].dt.year
-        df['Month'] = df['Date'].dt.month
-
-        # 첫 번째 월 구하기
-        first_month = df['Month'].iloc[0]
-        first_year = df['Year'].iloc[0]
-
-        # 각 월의 첫 거래일 찾기 (첫 번째 월은 제외)
-        monthly_data = []
-        for (year, month), group in df.groupby(['Year', 'Month']):
-            # 첫 번째 월 데이터는 건너뛰기
-            if year == first_year and month == first_month:
-                continue
-
-            # 월별 첫 날짜 선택
-            first_day = group.iloc[0]
-            monthly_data.append(first_day)
-
-        # 최종 tickvals 계산
-        if monthly_data:
-            monthly_df = pd.DataFrame(monthly_data)
-            tickvals = monthly_df["FormattedDate"].tolist()
-        else:
-            tickvals = []
-
-    # ✅ 모든 기간(1day, week, 1month, 1year)에서 캔들 차트 적용
-    fig.add_trace(go.Candlestick(
-        x=df["FormattedDate"],
-        open=df["Open"],
-        high=df["High"],
-        low=df["Low"],
-        close=df["Close"],
-        name="캔들 차트"
-    ))
-
-    fig.update_layout(
-        title=f"{company} 주가 ({period})",
-        xaxis_title="시간" if period == "1day" else "날짜",
-        yaxis_title="주가 (KRW)",
-        template="plotly_white",
-        xaxis=dict(
-            showgrid=True,
-            type="category",
-            tickmode='array',
-            tickvals=tickvals,
-            tickangle=-45
-        ),
-        hovermode="x unified"
-    )
-
-    st.plotly_chart(fig)
-
-
-    
-
-if __name__ == '__main__':
-    main()
